The DWC archive Timout was a problem loading the fungi archive.
So all /dwc requests make now a 302 redirect to a static version of the archive which are located in /static/dwc.

To regenrate the archive a pathon script is avalable build_curl_script.py. This should be run with cron.weekly or after a new list is
added. It regenerates all dwc archives using the /dwc_offline rest api.